added a system set for First schedule systems
