+++ src/lib.rs
@@ -8,13 +8,15 @@ mod camera_image_pipe;
 mod camera_node;
 mod camera_node_sobel;
 mod camera_readback;
+mod camera_strategy;
 mod plugin;
 mod widget;
 mod widget_halfblocks;
 mod widget_luminance;
 mod widget_none;
 
-pub use camera::{LuminanceConfig, RatatuiCamera, RatatuiCameraStrategy, RatatuiSubcamera};
+pub use camera::{RatatuiCamera, RatatuiCameraSet, RatatuiSubcamera};
 pub use camera_edge_detection::{EdgeCharacters, RatatuiCameraEdgeDetection};
+pub use camera_strategy::{LuminanceConfig, RatatuiCameraStrategy};
 pub use plugin::RatatuiCameraPlugin;
 pub use widget::RatatuiCameraWidget;
